requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==5.2.3
django-cleanup==9.0.0
django-crispy-forms==2.4
pillow==11.2.1
sqlparse==0.5.3

django-ckeditor~=6.7.3